frontend/requirements.txt
requests
typing
streamlit
python-dotenv